Now we will only create the orig file once.

+++ pocket_builder.sh
@@ -71,7 +71,9 @@ echo '*********************************************'
 # new template.  From there we need to replace the temnplated value with
 # the browser that we will be using.
 echo '[*] Replacing the default pocket-home config'
-sudo mv /usr/share/pocket-home/config.json /usr/share/pocket-home/config-orig.json
+if [ ! -f /usr/share/pocket-home/config-orig.json ];then
+	sudo mv /usr/share/pocket-home/config.json /usr/share/pocket-home/config-orig.json
+fi
 sudo wget -O /usr/share/pocket-home/config.json $REPO/configs/pocket-home-template.json
 sudo sed -i "s/BROWSER_EXEC/${BROWSER}/g" /usr/share/pocket-home/config.json
